+++ lib/swagger.js
@@ -19,26 +19,18 @@
 var _ = require('lodash');
 var allowedMethods = ['get', 'post', 'put', 'patch', 'delete'];
 var allowedDataTypes = ['string', 'integer', 'boolean', 'array'];
-var params = require(__dirname + '/paramTypes.js');
 var toJsonType = require('./toJsonType');
 var shallowClone = require('./shallowClone');
-
-// Stop express ressource with error code
-var stopWithError = function(res, error) {
-  this.setHeaders(res);
-  if (error && error.message && error.code)
-    res.send(JSON.stringify(error), error.code);
-  else
-    res.send(JSON.stringify({
-      'message': 'internal error',
-      'code': 500
-    }), 500);
-};
-
+var resourceHelpers = require('./resourceHelpers');
+var wrap = resourceHelpers.wrap;
+var appendToApi = resourceHelpers.appendToApi;
 
 // TODO-3.0.0 REMOVE
 var ignoreAppHandlerInConstructor = true;
 
+// TODO-3.0.0 REMOVE
+// For backwards compatability, we just export a new instance of Swagger
+module.exports = exports = new Swagger();
 
 function Swagger(appHandler) {
 
@@ -128,7 +120,7 @@ Swagger.prototype.setResourceListingPaths = function(app) {
       var r = self.resources[p] || self.resources[p.replace(self.formatString, '')];
       if (!r) {
         console.error('unable to find listing');
-        return stopWithError(res, {
+        return self.stopWithError(res, {
           'message': 'internal error',
           'code': 500
         });
@@ -200,7 +192,7 @@ Swagger.prototype.filterApiListing = function(req, res, r) {
   var excludedPaths = [];
 
   if (!r || !r.apis) {
-    return stopWithError(res, {
+    return self.stopWithError(res, {
       'message': 'internal error',
       'code': 500
     });
@@ -528,99 +520,21 @@ Swagger.prototype.addModels = function(models) {
   return this;
 };
 
-function wrap(callback, req, resp) {
-  callback(req, resp);
-}
-
-// appends a spec to an existing operation
-
-function appendToApi(rootResource, api, spec) {
-  var validationErrors = [];
-
-  if (!spec.nickname || spec.nickname.indexOf(' ') >= 0) {
-    //  nicknames don't allow spaces
-    validationErrors.push({
-      'path': api.path,
-      'error': 'invalid nickname "' + spec.nickname + '"'
-    });
-  }
-  // validate params
-  _.forOwn(spec.parameters, function (parameter) {
-
-    switch (parameter.paramType) {
-    case 'path':
-      if (api.path.indexOf('{' + parameter.name + '}') < 0) {
-        validationErrors.push({
-          'path': api.path,
-          'name': parameter.name,
-          'error': 'invalid path'
-        });
-      }
-      break;
-    case 'query':
-      break;
-    case 'body':
-      break;
-    case 'form':
-      break;
-    case 'header':
-      break;
-    default:
-      validationErrors.push({
-        'path': api.path,
-        'name': parameter.name,
-        'error': 'invalid param type ' + parameter.paramType
-      });
-      break;
-    }
-  });
-
-  if (validationErrors.length > 0) {
-    console.error(validationErrors);
-    return;
-  }
-
-  if (!api.operations) {
-    api.operations = [];
-  }
-
-  // TODO: replace if existing HTTP operation in same api path
-  var op = {
-    'parameters': spec.parameters,
-    'method': spec.method,
-    'notes': spec.notes,
-    'responseMessages': spec.responseMessages,
-    'nickname': spec.nickname,
-    'summary': spec.summary,
-    'consumes' : spec.consumes,
-    'produces' : spec.produces
-  };
-
-  // Add custom fields.
-  op = _.extend({}, spec, op);
-
-  if (!spec.type) {
-    op.type = 'void';
-  }
-  api.operations.push(op);
-
-  if (!rootResource.models) {
-    rootResource.models = {};
-  }
-}
-
 Swagger.prototype.addValidator = function(v) {
   this.validators.push(v);
 };
 
-// Create Error JSON by code and text
-
-function error(code, description) {
-  return {
-    'code': code,
-    'message': description
-  };
-}
+// Stop express ressource with error code
+Swagger.prototype.stopWithError = function(res, error) {
+  this.setHeaders(res);
+  if (error && error.message && error.code)
+    res.send(JSON.stringify(error), error.code);
+  else
+    res.send(JSON.stringify({
+      'message': 'internal error',
+      'code': 500
+    }), 500);
+};
 
 Swagger.prototype.setApiInfo = function(data) {
   this.apiInfo = data;
@@ -694,17 +608,3 @@ Swagger.prototype.configureDeclaration = function(resourceName, obj) {
     }
   }
 };
-
-// TODO-3.0.0 REMOVE
-// For backwards compatability, we just export a new instance of Swagger
-module.exports = exports = new Swagger();
-
-exports.params = params;
-exports.queryParam = exports.params.query;
-exports.pathParam = exports.params.path;
-exports.bodyParam = exports.params.body;
-exports.formParam = exports.params.form;
-exports.headerParam = exports.params.header;
-exports.error = error;
-exports.stopWithError = stopWithError;
-exports.stop = stopWithError;
